Provider - hanlde Tasks with URL support
