+++ examples/medplum-provider/src/pages/messages/MessagesPage.module.css
@@ -9,3 +9,15 @@
 .rightBorder {
   border-right: 1px solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-gray-8));
 }
+
+.button {
+  height: 32px;
+  background-color: light-dark(var(--mantine-color-gray-1), var(--mantine-color-gray-8));
+  color: light-dark(var(--mantine-color-black), var(--mantine-color-gray-4));
+  font-weight: 500;
+}
+
+.selected {
+  background-color: light-dark(var(--mantine-color-black), var(--mantine-color-gray-6));
+  color: light-dark(var(--mantine-color-white), var(--mantine-color-white));
+}
