+++ examples/medplum-provider/src/pages/tasks/TaskDetails.tsx
@@ -1,200 +1,82 @@
-import { Box, Button, Card, Grid, Modal, Stack, Text, Textarea } from '@mantine/core';
-import { notifications } from '@mantine/notifications';
-import { createReference, formatHumanName, getReferenceString, normalizeErrorString } from '@medplum/core';
-import { HumanName, Practitioner, Reference, Task } from '@medplum/fhirtypes';
-import { CodeInput, DateTimeInput, Loading, ResourceInput, useMedplum, useMedplumProfile } from '@medplum/react';
-import { IconCircleCheck, IconCircleOff } from '@tabler/icons-react';
-import { JSX, useEffect, useState } from 'react';
-import { Link, useNavigate, useParams } from 'react-router';
-import { usePatient } from '../../hooks/usePatient';
-import classes from './TaskDetails.module.css';
+import { Box, Flex, Paper, ScrollArea, SegmentedControl } from '@mantine/core';
+import { JSX, useState } from 'react';
+import { ResourceType, Task } from '@medplum/fhirtypes';
+import { PatientSummary, ResourceTimeline, useResource } from '@medplum/react';
+import { MedplumClient } from '@medplum/core';
+import { TaskInfo } from '../../components/tasks/TaskInfo';
+import { TasksInputNote } from '../../components/tasks/TaskInputNote';
+import { TaskSelectEmpty } from '../../components/tasks/TaskSelectEmpty';
+import classes from './TasksPage.module.css';
+import { useOutletContext } from 'react-router';
 
-export const TaskDetails = (): JSX.Element => {
-  const { patientId, encounterId, taskId } = useParams();
-  const patient = usePatient();
-  const medplum = useMedplum();
-  const navigate = useNavigate();
-  const author = useMedplumProfile();
-  const [task, setTask] = useState<Task | undefined>(undefined);
-  const [isOpened, setIsOpened] = useState(true);
-  const [practitioner, setPractitioner] = useState<Practitioner | undefined>();
-  const [dueDate, setDueDate] = useState<string | undefined>();
-  const [status, setStatus] = useState<Task['status'] | undefined>();
-  const [note, setNote] = useState<string>('');
+interface TasksOutletContext {
+  task: Task | undefined;
+  onTaskChange: (task: Task) => void;
+  onDeleteTask: (task: Task) => void;
+}
 
-  useEffect(() => {
-    const fetchTask = async (): Promise<void> => {
-      const task = await medplum.readResource('Task', taskId as string);
-      setStatus(task.status as typeof status);
-      setTask(task);
-      setDueDate(task.restriction?.period?.end);
-    };
+export function TaskDetails(): JSX.Element {
+  const { task, onTaskChange, onDeleteTask } = useOutletContext<TasksOutletContext>();
+  const selectedPatient = useResource(task?.for);
+  const [activeTab, setActiveTab] = useState<string>('properties');
 
-    fetchTask().catch((err) => {
-      notifications.show({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: normalizeErrorString(err),
-      });
-    });
-  }, [medplum, taskId]);
-
-  const handleOnSubmit = async (): Promise<void> => {
-    if (!task) {
-      return;
-    }
-
-    const updatedTask: Task = {
-      ...task,
-    };
-
-    const trimmedNote = note.trim();
-    if (trimmedNote !== '') {
-      updatedTask.note = [
-        ...(task.note || []),
-        {
-          text: trimmedNote,
-          authorReference: author && createReference(author),
-          time: new Date().toISOString(),
-        },
-      ];
-    }
-
-    if (status) {
-      updatedTask.status = status;
-    }
-
-    if (dueDate) {
-      updatedTask.restriction = {
-        ...updatedTask.restriction,
-        period: {
-          ...updatedTask.restriction?.period,
-          end: dueDate,
-        },
-      };
-    }
-
-    if (practitioner) {
-      updatedTask.owner = { reference: getReferenceString(practitioner) } as Reference<Practitioner>;
-    }
-
-    try {
-      await medplum.updateResource(updatedTask);
-      notifications.show({
-        icon: <IconCircleCheck />,
-        title: 'Success',
-        message: 'Task updated',
-      });
-      setTask(updatedTask);
-      navigate(`/Patient/${patientId}/Encounter/${encounterId}`)?.catch(console.error);
-    } catch {
-      notifications.show({
-        color: 'red',
-        icon: <IconCircleOff />,
-        title: 'Error',
-        message: 'Failed to update the task.',
-      });
-    }
+  const handleTabChange = (value: string): void => {
+    setActiveTab(value);
   };
 
   if (!task) {
-    return <Loading />;
+    return <TaskSelectEmpty />;
   }
 
   return (
-    <Modal
-      opened={isOpened}
-      onClose={() => {
-        navigate(-1)?.catch(console.error);
-        setIsOpened(false);
-      }}
-      size="xl"
-      styles={{
-        body: {
-          padding: 0,
-          height: '60vh',
-        },
-      }}
-    >
-      <Stack h="100%" justify="space-between" gap={0}>
-        <Box flex={1} miw={0}>
-          <Grid p="md" h="100%">
-            <Grid.Col span={6} pr="lg">
-              <Stack gap="sm">
-                <Card p="md" radius="md" className={classes.taskDetails}>
-                  <Stack gap="sm">
-                    <Text fz="lg" fw={700}>
-                      {task?.code?.text}
-                    </Text>
-                    {task?.description && <Text>{task.description}</Text>}
-                    {patient?.name && (
-                      <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
-                        <Text>View Patient</Text>
-                        <Button variant="subtle" component={Link} to={`/Patient/${patient.id}`}>
-                          {formatHumanName(patient.name?.[0] as HumanName)}
-                        </Button>
-                      </div>
-                    )}
-                  </Stack>
-                </Card>
+    <Flex direction="row" w="100%" h="100%">
+      <Flex direction="column" w="60%" h="100%" className={classes.borderRight}>
+        <TasksInputNote key={task.id} task={task} onDeleteTask={onDeleteTask} />
+      </Flex>
 
-                <ResourceInput
-                  name="practitioner"
-                  resourceType="Practitioner"
-                  label="Assigned to"
-                  defaultValue={task?.owner ? { reference: task.owner.reference } : undefined}
-                  onChange={(value) => {
-                    setPractitioner(value as Practitioner);
-                  }}
-                />
+      <Flex direction="column" w="40%" h="100%">
+        <Paper h="100%">
+          <Box px="md" pb="md" pt="md">
+            <SegmentedControl
+              value={activeTab}
+              onChange={(value: string) => handleTabChange(value)}
+              data={[
+                { label: 'Properties', value: 'properties' },
+                { label: 'Activity Log', value: 'activity-log' },
+                { label: 'Patient Summary', value: 'patient-summary' },
+              ]}
+              fullWidth
+              radius="md"
+              color="gray"
+              size="md"
+              className={classes.segmentedControl}
+            />
+          </Box>
 
-                <DateTimeInput
-                  name="Due Date"
-                  placeholder="End"
-                  label="Due Date"
-                  defaultValue={dueDate}
-                  onChange={setDueDate}
-                />
-
-                {task?.status && (
-                  <CodeInput
-                    name="status"
-                    label="Status"
-                    binding="http://hl7.org/fhir/ValueSet/task-status|4.0.1"
-                    maxValues={1}
-                    defaultValue={status}
-                    onChange={(value) => {
-                      if (value) {
-                        setStatus(value as typeof status);
-                      }
+          {selectedPatient?.resourceType === 'Patient' && task && (
+            <>
+              {activeTab === 'properties' && (
+                <TaskInfo p="md" key={task.id} task={task} onTaskChange={onTaskChange} />
+              )}
+              {activeTab === 'activity-log' && (
+                <ScrollArea h="calc(100% - 50px)">
+                  <ResourceTimeline
+                    value={task}
+                    loadTimelineResources={async (
+                      medplum: MedplumClient,
+                      _resourceType: ResourceType,
+                      id: string
+                    ) => {
+                      return Promise.allSettled([medplum.readHistory('Task', id)]);
                     }}
                   />
-                )}
-              </Stack>
-            </Grid.Col>
-
-            <Grid.Col span={6} pr="md">
-              <Stack gap="sm">
-                <Text>Note</Text>
-                <Text c="dimmed">Optional free form details about this task</Text>
-                <Textarea
-                  placeholder="Add note to this task"
-                  minRows={3}
-                  value={note}
-                  onChange={(event) => setNote(event.currentTarget.value)}
-                />
-              </Stack>
-            </Grid.Col>
-          </Grid>
-        </Box>
-
-        <Box className={classes.footer} h={70} p="md">
-          <Button variant="filled" onClick={handleOnSubmit}>
-            Save Changes
-          </Button>
-        </Box>
-      </Stack>
-    </Modal>
+                </ScrollArea>
+              )}
+              {activeTab === 'patient-summary' && <PatientSummary patient={selectedPatient} />}
+            </>
+          )}
+        </Paper>
+      </Flex>
+    </Flex>
   );
-};
+}
