+++ examples/medplum-provider/src/pages/tasks/TasksPage.module.css
@@ -31,3 +31,15 @@
   padding: var(--mantine-spacing-xs) var(--mantine-spacing-md);
   color: light-dark(var(--mantine-color-dark-9), var(--mantine-color-gray-0));
 }
+
+.button {
+  height: 32px;
+  background-color: light-dark(var(--mantine-color-gray-1), var(--mantine-color-gray-8));
+  color: light-dark(var(--mantine-color-black), var(--mantine-color-gray-4));
+  font-weight: 500;
+}
+
+.selected {
+  background-color: light-dark(var(--mantine-color-black), var(--mantine-color-gray-6));
+  color: light-dark(var(--mantine-color-white), var(--mantine-color-white));
+}
